Moving test case into correct location.
* See Issue #1298 for details.
* Fixes #1298
